Remove `ConstantMap<Rational>`
This is in preparation of having two different kinds (`CONST_RATIONAL`
and `CONST_INT`) share the same payload. To do so, we cannot rely on
`ConstantMap<Rational>` anymore to map the payload type to a kind. This
commit extends support in the `mkmetakind` script to deal with such
payloads by adding a `+` suffix to the type. The commit also does some
minor refactoring of `NodeManager::mkConst()` and
`NodeManager::mkConstInternal()` to support setting the kind explicitly.
Finally, the commit addresses all instances where `mkConst<Rational>()`
is used, including the API.