Faster hasing for `cvc5::String`
Previously, our hashing algorithm for `cvc5::String` would create an
`std::string` representation of the string and then call
`std::hash<std::string>` on the result. Creating the temporary
`std::string` is quite expensive and can be avoided by computing the
hash directly on the elements of the vector of characters in
`cvc5::String`.

This change improves our solving time on an industrial benchmark from
about 59s to 55s.